Flytter i18n ut av index.tsx
