+++ src/index.tsx
@@ -3,51 +3,22 @@ import "@navikt/ds-css";
 import {Suspense, useEffect} from "react";
 import {createRoot} from "react-dom/client";
 import {logWindowError} from "./lib/utils/loggerUtils";
-import {
-    DecoratorLocale,
-    injectDecoratorClientSide,
-    onLanguageSelect,
-    setAvailableLanguages,
-    setParams,
-} from "@navikt/nav-dekoratoren-moduler";
+import {injectDecoratorClientSide} from "@navikt/nav-dekoratoren-moduler";
 import {RouterProvider} from "react-router-dom";
 import {router} from "./routes";
 import {QueryClient, QueryClientProvider} from "@tanstack/react-query";
-import "./lib/i18n";
-
 import {ReactQueryDevtools} from "@tanstack/react-query-devtools";
-import i18n, {SUPPORTED_LANGUAGES} from "./lib/i18n";
-import {logAmplitudeEvent} from "./lib/utils/amplitude";
+import {i18nSetLangFromLocalStorage} from "./lib/i18n";
 import {withFaroProfiler} from "@grafana/faro-react";
-import {basePath} from "./lib/config";
 import {ApplicationSpinner} from "./lib/components/animasjoner/ApplicationSpinner";
 import {ValideringsContextProvider} from "./lib/valideringContextProvider";
 
 window.onerror = logWindowError;
+
 const queryClient = new QueryClient();
 
 const App = () => {
-    useEffect(() => {
-        setAvailableLanguages(
-            SUPPORTED_LANGUAGES.map((locale) => ({locale: locale, url: basePath, handleInApp: true}))
-        );
-
-        const handleLanguageSelect = (language: {locale: DecoratorLocale}) => {
-            i18n.changeLanguage(language.locale);
-            setParams({language: language.locale});
-            localStorage.setItem("digisos-language", language.locale);
-
-            logAmplitudeEvent("Valgt språk", {language: language.locale});
-        };
-
-        const storedLanguage = localStorage.getItem("digisos-language");
-        if (storedLanguage) {
-            i18n.changeLanguage(storedLanguage);
-            setParams({language: storedLanguage as DecoratorLocale});
-        }
-
-        onLanguageSelect(handleLanguageSelect);
-    }, []);
+    useEffect(i18nSetLangFromLocalStorage, []);
 
     return (
         <Suspense fallback={<ApplicationSpinner />}>
